+++ tests/test_metadata.py
@@ -1,39 +1,16 @@
-import os
 from pathlib import Path
-from typing import Callable
 
-import pytest
 from oyLabImaging import Metadata
 
-DATA = Path(__file__).parent / "data"
 
-
-@pytest.fixture
-def tmp_data_dir(tmp_path: Path) -> Callable[[str], str]:
-    """Return a function that creates a symlink to a file in the data dir.
-
-    Examples
-    --------
-    def test_something(tmp_data_dir):
-        path_to_file = tmp_data_dir("some_file_in_data_dir.nd2")
-    """
-
-    def _tmp_data_dir(filename: str):
-        dest = tmp_path / filename
-        os.symlink(DATA / filename, dest)
-        return str(dest)
-
-    return _tmp_data_dir
-
-
-def test_metadata(tmp_data_dir):
-    MD = Metadata(tmp_data_dir("sample_t3c2y32x32.nd2"))
+def test_metadata(t3c2y32x32):
+    MD = Metadata(t3c2y32x32)
     assert MD.type == "ND2"
     assert list(MD.channels) == ["Widefield Green", "Widefield Red"]
 
 
-def test_drift_correction(tmp_path, tmp_data_dir):
-    MD = Metadata(tmp_data_dir("sample_t3c2y32x32.nd2"))
+def test_drift_correction(tmp_path, t3c2y32x32):
+    MD = Metadata(t3c2y32x32)
     assert "driftTform" not in MD.image_table
     MD.CalculateDriftCorrection(Channel="Widefield Green", GPU=False)
     assert "driftTform" in MD.image_table
@@ -45,12 +22,3 @@ def test_drift_correction(tmp_path, tmp_data_dir):
     assert reloaded.type == "ND2"
     assert list(reloaded.channels) == ["Widefield Green", "Widefield Red"]
     assert "driftTform" in reloaded.image_table
-
-
-def test_metadata_viewer(tmp_data_dir):
-    napari = pytest.importorskip("napari")
-
-    MD = Metadata(tmp_data_dir("sample_t3c2y32x32.nd2"))
-    viewer = MD.viewer()
-    assert isinstance(viewer, napari.Viewer)
-    viewer.close()
